todo-list-application/todolist_api_django/api/requirements.txt
Django>=3.2,<3.2.8
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework>=3.11,<3.12
django-cors-headers>=3.7,<3.8
firebase-admin==5.0.2